+++ akashic-storage/src/SocketServer.ts
@@ -1,12 +1,16 @@
 import type { Server as HttpServer } from "node:http";
 import { Server, ServerOptions } from "socket.io";
 import { AMFlowServerManager } from "./AMFlowServerManager";
-import { initializeSocket } from "./initializeSocket";
+import {
+    initializeSocket,
+    InitializeSocketParameterObject,
+} from "./initializeSocket";
 
 interface SocketServerParameterObject {
     basePath: string;
     http: HttpServer;
     amfManager: AMFlowServerManager;
+    transfer: InitializeSocketParameterObject;
     /**
      * if undefined or empty array, skip setting cors.
      */
@@ -17,12 +21,14 @@ export class SocketServer {
     _basePath: string;
     _http: HttpServer;
     _amfManager: AMFlowServerManager;
+    _transfer: InitializeSocketParameterObject;
     _server: Server;
 
     constructor(param: SocketServerParameterObject) {
         this._basePath = param.basePath;
         this._http = param.http;
         this._amfManager = param.amfManager;
+        this._transfer = param.transfer;
         this._server = this._createServer(param.allowOrigins);
     }
 
@@ -40,7 +46,7 @@ export class SocketServer {
         }
         const io = new Server(this._http, opts);
         io.on("connection", (socket) => {
-            initializeSocket(socket, this._amfManager);
+            initializeSocket(socket, this._amfManager, this._transfer);
         });
         return io;
     }